test(plugins): tag the store tests that need POSIX behavior
Two bundled-store tests assert what only a Unix filesystem does: a
directory mode denying writes to its owner, and rename(2) reporting
ENOTDIR or ENOTEMPTY rather than replacing foreign data. They belong
beside the package's other build-tagged test support rather than in the
portable file.
